Modified LogsWriter, Added logings to the Tests
for LogsWriter -> changed the file name of the logs, cause was not
according to it's task. Old: Logs.log; New: TestingLog.log.
for Tests -> added a logging Systems that works the way now follows:
When a method in a Junit test is called to test some specific part of
the class that's being tested there is an indicator, it is also logged,
of the start of the method who makes some asserts and at the end if the
asserts are passed then it log other message saying it was passed.

In this pack is also being uploaded a copy of the files of logs I've
created.